// src/lib/managers/sonarr/generated.ts
export interface paths {
    "/api": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/api/v3/series": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: {
                    tvdbId?: number;
                    includeSeasonImages?: boolean;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["SeriesResource"][];
                    };
                };
            };
        };
        put?: never;
        post: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: {
                content: {
                    "application/json": components["schemas"]["SeriesResource"];
                };
            };
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["SeriesResource"];
                    };
                };
            };
        };
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v3/series/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: {
            parameters: {
                query?: {
                    includeSeasonImages?: boolean;
                };
                header?: never;
                path: {
                    id: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["SeriesResource"];
                    };
                };
            };
        };
        put: {
            parameters: {
                query?: {
                    moveFiles?: boolean;
                };
                header?: never;
                path: {
                    id: string;
                };
                cookie?: never;
            };
            requestBody?: {
                content: {
                    "application/json": components["schemas"]["SeriesResource"];
                };
            };
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["SeriesResource"];
                    };
                };
            };
        };
        post?: never;
        delete: {
            parameters: {
                query?: {
                    deleteFiles?: boolean;
                    addImportListExclusion?: boolean;
                };
                header?: never;
                path: {
                    id: number;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
// ...
        SeriesResource: {
            /** Format: int32 */
            id?: number;
            title?: string | null;
            alternateTitles?: components["schemas"]["AlternateTitleResource"][] | null;
            sortTitle?: string | null;
            status?: components["schemas"]["SeriesStatusType"];
            readonly ended?: boolean;
            profileName?: string | null;
            overview?: string | null;
            /** Format: date-time */
            nextAiring?: string | null;
            /** Format: date-time */
            previousAiring?: string | null;
            network?: string | null;
            airTime?: string | null;
            images?: components["schemas"]["MediaCover"][] | null;
            originalLanguage?: components["schemas"]["Language"];
            remotePoster?: string | null;
            seasons?: components["schemas"]["SeasonResource"][] | null;
            /** Format: int32 */
            year?: number;
            path?: string | null;
            /** Format: int32 */
            qualityProfileId?: number;
            seasonFolder?: boolean;
            monitored?: boolean;
            monitorNewItems?: components["schemas"]["NewItemMonitorTypes"];
            useSceneNumbering?: boolean;
            /** Format: int32 */
            runtime?: number;
            /** Format: int32 */
            tvdbId?: number;
            /** Format: int32 */
            tvRageId?: number;
            /** Format: int32 */
            tvMazeId?: number;
            /** Format: int32 */
            tmdbId?: number;
            /** Format: date-time */
            firstAired?: string | null;
            /** Format: date-time */
            lastAired?: string | null;
            seriesType?: components["schemas"]["SeriesTypes"];
            cleanTitle?: string | null;
            imdbId?: string | null;
            titleSlug?: string | null;
            rootFolderPath?: string | null;
            folder?: string | null;
            certification?: string | null;
            genres?: string[] | null;
            tags?: number[] | null;
            /** Format: date-time */
            added?: string;
            addOptions?: components["schemas"]["AddSeriesOptions"];
            ratings?: components["schemas"]["Ratings"];
            statistics?: components["schemas"]["SeriesStatisticsResource"];
            episodesChanged?: boolean | null;
            /**
             * Format: int32
             * @deprecated
             */
            readonly languageProfileId?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}